tracklist/requirements.txt
fastapi
fastapi[standard]
sqlmodel
pydantic-settings
psycopg2
pyjwt
